misc/HtmlTestSuite/results/test-0341.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>'background-color' property</title>
<link href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-get" rel="help"/>
<link href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-set" rel="help"/>
<link href="https://drafts.css-houdini.org/css-typed-om-1/#property-stle-value-normalization" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="../../resources/testhelper.js"/>
<script src="resources/testsuite.js"/>
</head><body>
<div id="log"/>
<script>
'use strict';

runPropertyTests('background-color', [
  {
    syntax: 'currentcolor',
    // computes to a &lt;color&gt;, which is not supported in level 1
    computed: (_, result) =&gt; assert_class_string(result, 'CSSStyleValue')
  }
]);

// &lt;color&gt;s are not supported in level 1
runUnsupportedPropertyTests('background-color', [
  'red', '#bbff00', 'rgb(255, 255, 128)', 'hsl(50, 33%, 25%)',
  'transparent'
]);

</script>
</body></html>
